requirements/base.txt
pytz==2021.1  # https://github.com/stub42/pytz
python-slugify==5.0.2  # https://github.com/un33k/python-slugify
Pillow==8.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==5.2.0  # https://github.com/evansd/whitenoise
redis==3.5.3  # https://github.com/andymccurdy/redis-py
hiredis==1.1.0  # https://github.com/redis/hiredis-py
celery==4.4.6  # pyup: < 5.0,!=4.4.7  # https://github.com/celery/celery
django-celery-beat==2.2.0  # https://github.com/celery/django-celery-beat
flower==0.9.7  # https://github.com/mher/flower


# Django
# ------------------------------------------------------------------------------
django==3.1.13  # pyup: < 3.2  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==4.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.44.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.12.0  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==5.0.0  # https://github.com/jazzband/django-redis
# Django REST Framework
djangorestframework==3.12.4  # https://github.com/encode/django-rest-framework
django-cors-headers==3.7.0 # https://github.com/adamchainz/django-cors-headers
django-jazzmin>=2.4.7,<3.0.0
django-guardian>=2.4.0,<3.0.0
djangorestframework-guardian>=0.3.0,<1.0.0
django-filter==2.4.0
crispy-bootstrap5>=0.4,<1.0
django-simple-history>=3.0.0,<4.0.0
django-admin-sortable>=2.2.4,<3.0.0
django-admin-sortable2>=1.0,<2.0
django-baton>=2.2.2,<3.0.0
djangorestframework-simplejwt>=4.7.2,<5.0.0
#dj-rest-auth>=2.1.11,<3.0.0
wagtail>=2.14.1,<3.0.0
wagtail-modelchooser>=2.13.1,<3.0.0
python-memcached>=1.59,<2.0
django-redis>=5.0.0,<6.0.0
django-cacheops>=6.0,<7.0
djoser>=2.1.0,<3.0.0
djangorestframework_simplejwt
Unidecode>=1.3.2,<2.0.0
django-meta>=2.0.0,<3.0.0
django-json-ld>=0.0.4,<1.0.0
drf-writable-nested>=0.6.4,<1.0.0

openpyxl>=3.2.0b1,<4.0.0